refactor SRQ AST
